+++ implementation/js/src/infrastructure/application.js
@@ -173,6 +173,8 @@ _.each(['Core', 'Util'], function(coreModule){
 		var $body = $('body');
 		function trackScreenSize(e, silent){
 			var screenSize = {h: $window.height(), w: $window.width()};
+			if(!validScreenSize(screenSize)) return;
+
 			////////////////cache the screen size/////////////
 			Application.screenSize = screenSize;
 			//////////////////////////////////////////////////
@@ -183,7 +185,16 @@ _.each(['Core', 'Util'], function(coreModule){
 			if(!silent)
 				Application.trigger('app:resized', screenSize);
 		}
+		function validScreenSize(size){
+			return size.h > 0 && size.w > 0;
+		}
 		$window.on('resize', _.debounce(trackScreenSize, Application.config.rapidEventDelay));
+		//check screen size, trigger app:resized and get app.screenSize ready.
+		Application._ensureScreenSize = function(done){
+			trackScreenSize(); 
+			if(!Application.screenSize) _.delay(Application._ensureScreenSize, Application.config.rapidEventDelay/4, done);
+			else done();
+		};
 
 		//Track window scroll
 		function trackScroll(){
@@ -335,12 +346,7 @@ _.each(['Core', 'Util'], function(coreModule){
 				console.error('DEV:Context-Guard-Error:', ctxName, error);
 			};			
 
-		//5 Check screen size BEFORE running all the initializers user has defined
-		Application.on('initialize:before', function(){
-			//check screen size, trigger app:resized and get app.screenSize ready.
-			trackScreenSize();
-		});
-		//6 Activate Routing AFTER running all the initializers user has defined
+		//5 Activate Routing AFTER running all the initializers user has defined
 		//Context Switching by Routes (can use href = #navigate/... to trigger them)
 		Application.on("initialize:after", function(options){
 			//init client page router and history:
@@ -350,7 +356,7 @@ _.each(['Core', 'Util'], function(coreModule){
 				},
 				controller: {
 					navigateTo: function(path){
-						Application.trigger('app:navigate', path || Application.config.defaultContext, true); //will skip updating #hash since the router is triggered by #hash change.
+						Application.navigate(path || Application.config.defaultContext, true); //will skip updating #hash since the router is triggered by #hash change.
 					},
 				}
 			});
@@ -408,11 +414,13 @@ _.each(['Core', 'Util'], function(coreModule){
 				if(!Application.Core.Context.get(Application.config.defaultContext))
 					console.warn('DEV::Application::You might want to define a Default context using app.create(\'Context Name\', {...})');
 				else
-					window.location.hash = ['#navigate', Application.config.defaultContext].join('/');
+					Application.navigate(Application.config.defaultContext);
 			}
 
 			//4. Start the app --> pre init --> initializers --> post init(router setup)
-			Application.start();
+			Application._ensureScreenSize(function(){
+				Application.start();
+			});
 
 		}