+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Change Log
 7. Fixed .circle() less mixin;
 8. Fixed app:resized event timing during app bootstrapping;
 9. app.mainView can now be overridden during 'app:before-template-ready' event;
+10. Ensured app.screenSize before app.start(); (available in all initializers)
 
 In progress...
 x. +User and api authorization to devserver